DD-60-fix: 타입과 데이터 연결 수정, 컴포넌트분리
